+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - New component, Logout, which logs out the current user when rendered.
 - New route, /reset-password, which renders the new Logout component if a user is already authenticated.
 - Display capacity/registrants/waitlist for each camp under Admin -> Camp Sessions.
+- Survey Question component to first time wizard
 
 ### Changed